// src/features/outputs/qa/types.ts
/*
Copyright IBM Corporation 2023

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

  http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

export interface IQuestion {
    id: string;
    type: string;
    description: string;
    hints?: Array<string>;
    options?: Array<string>;
    default?: unknown;
    answer?: unknown;
}

export interface IQAStep {
    question: IQuestion;
}

export interface IQASubComponentProps {
    isDisabled?: boolean;
    question: IQuestion;
    setAnswer: (a: unknown) => void;
}